refactor: Move banners to separate component
