+++ src/shared/components/Banners/EHRBanners/EHRBanners.const.ts
@@ -1,3 +1,4 @@
-// TODO: Replace with actual links before merging PR
+// TODO: Replace with actual links when microsite is ready
+// https://mindlogger.atlassian.net/browse/M2-8823
 export const EHR_LEARN_MORE_URL = 'https://mindlogger.org/';
 export const EHR_CONTACT_URL = 'https://mindlogger.org/';
