+++ src/shared/components/Banners/EHRBanners/index.ts
@@ -1,2 +1,3 @@
+export * from './EHRBanners';
 export * from './EHRBannerAvailable';
 export * from './EHRBannerActive';
